PortalBackend/Python/django/requirements.txt
Django==4.0.3
django-cors-headers==3.13.0
django-filter==22.1
django-jsonfield==1.4.1
djangorestframework==3.14.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
pytest==7.3.2
pytest-django==4.5.2
requests==2.31.0
